+++ Source/CesiumRuntime/Private/CesiumGltfComponent.h
@@ -66,7 +66,9 @@ public:
 
   static TUniquePtr<HalfConstructed> CreateOffGameThread(
       const glm::dmat4x4& Transform,
-      const CreateGltfOptions::CreateModelOptions& Options);
+      const CreateGltfOptions::CreateModelOptions& Options,
+      const CesiumGeospatial::Ellipsoid& Ellipsoid =
+          CesiumGeospatial::Ellipsoid::WGS84);
 
   static UCesiumGltfComponent* CreateOnGameThread(
       CesiumGltf::Model& model,